Cập nhật EventServiceImpl và EventController hoàn chỉnh

+++ src/main/java/com/example/uniclub/dto/request/WalletAdjustRequest.java
@@ -4,6 +4,6 @@ import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 
 public record WalletAdjustRequest(
-        @NotNull @Min(1) Integer amount,
+        @NotNull @Min(1) Long amount,
         String description
 ) {}
