+++ src/main/java/com/example/uniclub/service/EventWalletService.java
@@ -5,7 +5,7 @@ import com.example.uniclub.entity.Event;
 
 public interface EventWalletService {
     void createEventWallet(Event event);
-    void grantBudgetToEvent(Event event, int points);
+    void grantBudgetToEvent(Event event, long points);
     void returnSurplusToClubs(Event event);
     EventWalletResponse getEventWalletDetail(Long eventId);
