+++ src/main/java/com/example/uniclub/service/RewardService.java
@@ -1,7 +1,7 @@
 package com.example.uniclub.service;
 
 public interface RewardService {
-    void sendCheckInRewardEmail(Long userId, String eventName, int pointsEarned, int totalPoints);
-    void sendManualBonusEmail(Long userId, int bonusPoints, String reason, int totalPoints);
-    void sendMilestoneEmail(Long userId, int milestone);
+    void sendCheckInRewardEmail(Long userId, String eventName, long pointsEarned, long totalPoints);
+    void sendManualBonusEmail(Long userId, long bonusPoints, String reason, long totalPoints);
+    void sendMilestoneEmail(Long userId, long milestone);
 }
